[EG-440] Renaming of various parts of the error tool
