+++ settings.gradle
@@ -108,5 +108,5 @@ include 'serialization-deterministic'
 
 include 'tools:checkpoint-agent'
 include 'detekt-plugins'
-include 'tools:error-page-builder'
+include 'tools:error-tool'
