+++ tools/error-tool/src/main/resources/ErrorPageHeadings.properties
