+++ tools/error-tool/src/test/resources/ErrorPageHeadings.properties
